+++ src/test/java/li/cil/sedna/riscv/ISATests.java
@@ -146,7 +146,7 @@ public final class ISATests {
         }
 
         @Override
-        public int load(final int offset, final int sizeLog2) {
+        public long load(final int offset, final int sizeLog2) {
             assert sizeLog2 == Sizes.SIZE_32_LOG2;
             switch (offset) {
                 case 0x00: {
@@ -168,7 +168,7 @@ public final class ISATests {
         }
 
         @Override
-        public void store(final int offset, final int value, final int sizeLog2) {
+        public void store(final int offset, final long value, final int sizeLog2) {
             assert sizeLog2 == Sizes.SIZE_32_LOG2;
             switch (offset) {
                 case 0x00: {
